fix(registry): route unregistered one-shot workdirs at the ephemeral index dir (TRA-1136)
TRA-396 gave one-shot agent-run checkouts their own index subdirectory so
`sweepEphemeralDbs` could collect them by age. But the routing decision lived
only in `registerProject`, and an ephemeral root's registry entry is
process-local (`_ephemeralEntries`) — so every other process, and every code
path that never registers at all (notably `ProjectManager.addProject` with
`persist: false`), fell through to a bare `getDbPath()` and created the DB
straight in `INDEX_DIR`. Nothing reclaims that: the sweep only walks
`EPHEMERAL_INDEX_DIR`, and `prune` can only class it `orphan_unregistered`.

Confirmed on the reported machine: 42 of the 250 orphan DBs hash back to real
`multica_workspaces_*/<ws>/<run>/workdir/<repo>` paths, several created the
same day — i.e. roots `isEphemeralProjectRoot` already recognizes.

Fix: `getDbPath` picks the directory from the path shape itself, so the two
agree without a registry lookup. `isEphemeralProjectRoot` and its two patterns
move to global.ts (which must stay import-free) and are re-exported from
registry.ts, where every caller already imports them; `getEphemeralDbPath` and
`registerProject`'s ephemeral branch are now redundant and gone.

Agent: Lead Engineer (trace-mcp)
